VkSaveAlbumRenamer/requirements.txt
beautifulsoup4==4.12.3
easyargs==0.9.4
piexif==1.1.3
Pillow==11.0.0
python_dateutil==2.9.0.post0
